Move Ubw32 specific MessageProcessor functionality to separate class.
