+++ core/src/main/java/ch/desm/middleware/app/core/component/ComponentMessageProcessorBase.java
@@ -29,58 +29,10 @@ public abstract class ComponentMessageProcessorBase<T1> extends MessageProcessor
      */
     public void processEndpointMessage(ComponentBrokerClientBase component, String message, String topic) {
         if (message != null && !message.isEmpty()) {
-
             LOGGER.log(Level.INFO, "processing endpoint message: " + message + ", component: " + component + ", topic: " +topic);
             component.publish(message, topic);
         }else{
             LOGGER.log(Level.TRACE, "skipped processing endpoint message.");
-        }
-    }
-
-    // TODO: move to some place where it belongs to UBW32 endpoints only!
-    public void delegateToEndpoint(EndpointUbw32 endpoint, ComponentMapBase mapDigital, ComponentMapBase mapAnalog, String key, MessageCommon message){
-        switch(message.getType()) {
-            case BOOLEAN: {
-                // TODO: get rid of the maps?
-                if (!mapDigital.isKeyAvailable(key)) {
-                    throw new RuntimeException("Digital pin must be controlled through message of type boolean!");
-                }
-
-                final String endpointRegister = mapDigital.getValueForKey(key);
-                final String registerName = String.valueOf(endpointRegister.charAt(0));
-                final Integer pin = Integer.valueOf(endpointRegister.substring(1));
-
-                if (message.isInputMessage()) {
-                    throw new RuntimeException("Message should not trigger an input on the UBW32");
-                    //endpoint.getPinInputDigital(registerName, pin);
-                }
-
-                try {
-                    endpoint.setPinOutputDigital(registerName, pin, message.getParameterAsBoolean());
-                } catch (MessageCommon.BadParameterTypeCastException e) {
-                    // should not happen as we checked for bool type above already
-                }
-                break;
-            }
-            case DOUBLE: {
-                // TODO: get rid of the maps?
-                if (!mapAnalog.isKeyAvailable(key)) {
-                    throw new RuntimeException("Analog pin must be controlled through message of type double!");
-                }
-
-                final String endpointRegister = mapAnalog.getValueForKey(key);
-
-                if (message.isInputMessage()) {
-                    throw new RuntimeException("Message should not trigger an input on the UBW32");
-                    // endpoint.getPinInputAnalog(endpointRegister);
-                }
-
-                // TODO: implement through PWM
-                throw new RuntimeException("setting analog pins not implemented yet");
-                //break;
-            }
-            default:
-                throw new RuntimeException("Unable to map message of type " + message.getTypeName() + " to UBW32 pin");
         }
     }
 }
